+++ signify/authenticode/__init__.py
@@ -1,6 +1,5 @@
 from .authroot import AUTHROOTSTL_PATH, CertificateTrustList, CertificateTrustSubject
 from .signed_file import AuthenticodeFile
-from .signed_pe import SignedPEFile, SignedPEFingerprinter
 from .structures import (
     CERTIFICATE_LOCATION,
     TRUSTED_CERTIFICATE_STORE,
@@ -34,15 +33,5 @@ __all__ = [
     "RFC3161SignedData",
     "RFC3161SignerInfo",
     "SigInfo",
-    "SignedPEFile",
-    "SignedPEFingerprinter",
     "TSTInfo",
 ]
-
-# SignedMsiFile is not necessarily available, as olefile is an optional dependency.
-try:
-    from .signed_msi import SignedMsiFile
-except ImportError:
-    pass
-else:
-    __all__.extend(["SignedMsiFile"])