+++ tests/test_context.py
@@ -6,7 +6,7 @@ from signify.authenticode import (
     TRUSTED_CERTIFICATE_STORE,
     TRUSTED_CERTIFICATE_STORE_NO_CTL,
 )
-from signify.authenticode.signed_pe import SignedPEFile
+from signify.authenticode.signed_file.pe import SignedPEFile
 from signify.exceptions import VerificationError
 from signify.x509.certificates import Certificate
 from signify.x509.context import FileSystemCertificateStore, VerificationContext